requirements.txt
python-dotenv == 1.0.1
flask >= 3.0.0
psycopg[binary]  >= 3.0.0